+++ tests/action_diffs.rs
@@ -109,7 +109,7 @@ fn assert_action_diff_received(app: &mut App, action_diff_event: ActionDiffEvent
             assert_eq!(action_state.value(&action), 0.);
             assert!(action_state.axis_pair(&action).is_none());
         }
-        ActionDiff::ValueChanged { action, value } => {
+        ActionDiff::AxisChanged { action, value } => {
             assert!(action_state.pressed(&action));
             assert_eq!(action_state.value(&action), value);
         }
@@ -155,7 +155,7 @@ fn generate_binary_action_diffs() {
             ActionDiff::Released { .. } => {
                 panic!("Expected a `Pressed` variant got a `Released` variant")
             }
-            ActionDiff::ValueChanged { .. } => {
+            ActionDiff::AxisChanged { .. } => {
                 panic!("Expected a `Pressed` variant got a `ValueChanged` variant")
             }
             ActionDiff::AxisPairChanged { .. } => {
@@ -180,7 +180,7 @@ fn generate_binary_action_diffs() {
             ActionDiff::Pressed { .. } => {
                 panic!("Expected a `Released` variant got a `Pressed` variant")
             }
-            ActionDiff::ValueChanged { .. } => {
+            ActionDiff::AxisChanged { .. } => {
                 panic!("Expected a `Released` variant got a `ValueChanged` variant")
             }
             ActionDiff::AxisPairChanged { .. } => {
@@ -216,7 +216,7 @@ fn generate_value_action_diffs() {
         assert_eq!(action_diff_event.owner, Some(entity));
         assert_eq!(action_diff_event.action_diffs.len(), 1);
         match action_diff_event.action_diffs.first().unwrap().clone() {
-            ActionDiff::ValueChanged { action, value } => {
+            ActionDiff::AxisChanged { action, value } => {
                 assert_eq!(action, Action::PayTheBills);
                 assert_eq!(value, input_value);
             }
@@ -248,7 +248,7 @@ fn generate_value_action_diffs() {
             ActionDiff::Pressed { .. } => {
                 panic!("Expected a `Released` variant got a `Pressed` variant")
             }
-            ActionDiff::ValueChanged { .. } => {
+            ActionDiff::AxisChanged { .. } => {
                 panic!("Expected a `Released` variant got a `ValueChanged` variant")
             }
             ActionDiff::AxisPairChanged { .. } => {
@@ -294,7 +294,7 @@ fn generate_axis_action_diffs() {
             ActionDiff::Pressed { .. } => {
                 panic!("Expected a `AxisPairChanged` variant got a `Pressed` variant")
             }
-            ActionDiff::ValueChanged { .. } => {
+            ActionDiff::AxisChanged { .. } => {
                 panic!("Expected a `AxisPairChanged` variant got a `ValueChanged` variant")
             }
         }
@@ -316,7 +316,7 @@ fn generate_axis_action_diffs() {
             ActionDiff::Pressed { .. } => {
                 panic!("Expected a `Released` variant got a `Pressed` variant")
             }
-            ActionDiff::ValueChanged { .. } => {
+            ActionDiff::AxisChanged { .. } => {
                 panic!("Expected a `Released` variant got a `ValueChanged` variant")
             }
             ActionDiff::AxisPairChanged { .. } => {
@@ -371,7 +371,7 @@ fn process_value_action_diff() {
 
     let action_diff_event = ActionDiffEvent {
         owner: Some(entity),
-        action_diffs: vec![ActionDiff::ValueChanged {
+        action_diffs: vec![ActionDiff::AxisChanged {
             action: Action::PayTheBills,
             value: 0.5,
         }],